pyproject.toml
[tool.poetry]
name = "ngm"
version = "0.1.0"
description = "Nepal Government Modernization - Judicial Data Scrapers"
authors = ["Jawafdehi Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.14.0"
beautifulsoup4 = "^4.14.3"
python-dateutil = "^2.9.0.post0"
nepali = "^1.1.3"
boto3 = "^1.35.0"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
